+++ src/main/java/io/kaitai/struct/RandomAccessFileKaitaiStream.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2015-2019 Kaitai Project: MIT license
+ * Copyright 2015-2020 Kaitai Project: MIT license
  *
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
@@ -85,10 +85,9 @@ public class RandomAccessFileKaitaiStream extends KaitaiStream {
     }
 
     @Override
-    public int pos() {
+    public long pos() {
         try {
-            // FIXME cast
-            return (int) raf.getFilePointer();
+            return raf.getFilePointer();
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
